+++ src/components/Popups/Alerts/components/AlertFooter.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 
 import { AlertFooterProps } from "../types";
 
@@ -6,11 +6,26 @@ export function AlertFooter<TData = undefined>({
   data,
   confirmBtnText = "Confirm",
   confirmBtnDisabled = false,
+  disableConfirmOnEnter = false,
   onConfirm,
   declineBtnText = "Close",
   declineBtnDisabled = false,
   onDecline,
 }: AlertFooterProps<TData>) {
+  useEffect(() => {
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (onConfirm && event.code === "Enter" && !disableConfirmOnEnter) {
+        onConfirm(data as unknown as TData);
+      }
+    };
+
+    document.addEventListener("keydown", handleKeyDown);
+
+    return () => {
+      document.removeEventListener("keydown", handleKeyDown);
+    };
+  }, []);
+
   if (!onConfirm && !onDecline) {
     return null;
   }